Update BASE_IMAGE_TAG
In the Dockerfile, we've moved down to php7.4-apache, but in here there was still a reference to the 8.1 image and the build error says "imagestreamtag.image.openshift.io/php:8.1-apache configured" so there seems to be a mismatch here, though I don't really see how updating it in an echo is going to fix anything? I guess the output of the echo is used as the basis of the command? Probably not using the right words here